fix: typo and minor errors (#13)
* Add NOT NULL contrsint to the hints in 140

The NOT NULL constraint was almost ignored from all the hints and from
the final solution. Fixed it.

* Fix a minor typo in 190

* Fix the hint 200 by adding appropriate keywords to be used

There is no need to use SERIAL and PRIMARY key for a VARCHAR name
column. Use VARCHAR() & NOT NULL instead.

* Correct the 220 statement

* Change variable to query the database

We were supposed to use BIKE_ID_TO_RETURN variable and not the
BIKE_ID_TO_RENT to get the RENT_ID

* fix:Update tutorial.json